requirements.txt
aiogram==2.21
selenium==4.10.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0